correction
git-svn-id: svn://elaine/murooma/trunk@162 d8a302eb-03bc-478d-80e4-98257eca68ef

+++ stream.h
@@ -27,7 +27,8 @@ private:
 	void updateBuffers(int age);
 	void resetBuffers();
 
-	int sleep;
+	float sleep;
+//	int correction;
 	Queue<std::shared_ptr<Chunk>> chunks;
 	DoubleBuffer<int>* pMiniBuffer;
 	DoubleBuffer<int>* pBuffer;
